Fill in gzip stubs for resource tools library
In terms of values passed to the compression library, the functionality is as similar to the Python gzip library as possible. However, when compressing files, the header of the compressed data does not exactly match that of the files generated from Python. Leaving this as is for now because the header data for the generated resource files contains the filename, and that is currently not a part of the resource tools interface.