Start filling CLI implementation & support destination Remote-CDN
WIP commit for collaboration purposes.

CLI interface and command implementation begun for patch & bundle creation.
Added missing Remote-CDN destination, this destination type creates
compressed files that are ready to be uploaded to the CDN.
Experimental EVE style patch creation tests have been added but commented out
this is a convenience measure, they contain absolute paths which are nonesense and
require specific files available on the local machine to run. The tests do not
call initialise on resource-tools so download will fail, this can work with a hack
but the issue will go away once the RAII download PR is merged.