Fix streaming compression
Make sure we don't pass the finish flag when insufficient data exists in the buffer, rather wait for the user to call Finish. Fix crash due to trying to return a string the size of all the data processed so far, when we have less data than this on hand.