Fix issue when compressing files
The old approach was relying on the >> operator of BundleStreamOut, which would output a chunk of the data in m_cache. However, recent changes have introduced some changes involving compressing the data. Since all is required is to read the data from the buffer, rely solely on the ReadBytes method.