Improve bundle algorithm and target toolset v141
Bundle slightly improved to deal better with large files.
Still work to be done to fully support bundling of large files, details in TODO.

Project now builds on v141 so that our agents can build the project.
Some changes were required to get cli working as win compiler was
encountering an internal error trying to process argparse.
Argparse version pushed back to time prior to the problem code being added.
Changes made to accomadate for functionality that was being used and was
added after the version now pinned to. Specifically subparsers.
When we move to v143 it would be good to reintroduce subparsers as they
produce better commandline help.