Fix problems when bundling Patch Resource Groups
Fix an error where the bundling would fail to load the parts of the patch that don't have a file associated with them. (These resources get sourced directly from the source file at the specified offset)

Also, make sure that the Resource Group that is required in order to apply a patch gets included in the bundle.