Set runtime to MT rather than MD
This is a requirement to meet how launcher would like to ingest the library.

The change in CcpBuildConfigurations.cmake is probably most notable.
This was stopping correct building when using MT.
The reason this was added in the first place is unknown, p4 history only
states that it was added with no further information.
It is likely that this setting is necessary for something, however as this
file is no longer global to all components, changing it here only has
an effect on carbon-resources so fallout is unlikely.